refactor: button common component
